pyproject.toml
[project]
name = "openLMS"
version = "0.1.0"
description = "OpenLMS is a laundry management system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "celery>=5.5.3",
    "crispy-bootstrap5>=2025.6",
    "dj-database-url>=3.0.1",
    "django>=4.2.0,<5.0",
    "django-allauth>=65.10.0",
    "django-cors-headers>=4.7.0",
    "django-crispy-forms>=2.4",
    "django-debug-toolbar>=6.0.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-guardian>=3.0.3",
    "django-htmx>=1.23.2",
    "django-storages>=1.14.6",
    "django-widget-tweaks>=1.5.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "factory-boy>=3.3.3",
    "flake8>=7.3.0",
    "gunicorn>=23.0.0",
    "isort>=6.0.1",
    "model-bakery>=1.20.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "phonenumbers>=9.0.10",
    "pillow>=11.3.0",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
    "pytz>=2025.2",
    "redis>=6.2.0",
    "reportlab>=4.4.3",
    "sentry-sdk>=2.33.2",
    "whitenoise>=6.9.0",
    "xlsxwriter>=3.2.5",
]
